contributors/ttwf_bj/tang,shaofeng/incoming/border-image-repeat_repeatnegx_none_50px.html

<html>
<head>
    <title> CSS3 Backgroud Border Test: "border-image-repeat:repeat-x;height:200px;width:200px;border-image-source:none;border-image-width:50px" on test div</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-css3-background-20120214/#the-border-image-repeat" />
    <meta name="assert" content="Check if 'border-image-repeat:repeat-x;height:200px;width:200px;border-image-source:none;border-image-width:50px' work on div" />
    <script type="text/javascript" src="../resources/testharness.js"></script>
    <script type="text/javascript" src="../resources/testharnessreport.js"></script>
    <script type="text/javascript" src="support/cssstyle.js"></script>
    <style>
        #test{
            height: 30px;
            width: 80px;
            padding: 5px;
            border: 5px solid black;
            margin: 5px;
            background: blue;
        }
    </style>
</head>
<body>
    <div id="log"></div>
    <div id="test"></div>
    <script type="text/javascript">
        var div = document.querySelector("#test");
        var t = async_test(document.title, { timeout: 500 });
        t.step(function () {
            div.style[headProp("border-image-repeat")] = "repeat-x";
            div.style[headProp("height")] = "200px";
            div.style[headProp("width")] = "200px";
            div.style[headProp("border-image-source")] = "none";
            div.style[headProp("border-image-width")] = "50px";
            var propvalue = GetCurrentStyle("border-image-repeat");
            var prop = propvalue.indexOf("repeat-x")!=-1;
            assert_false(prop, "The element border-image-repeat can be repeat-x")
            var height = GetCurrentStyle("height");
            prop = height.indexOf("200px")!=-1;
            assert_true(prop, "The element height should be 200px");
            var width = GetCurrentStyle("width");
            prop = width.indexOf("200px")!=-1;
            assert_true(prop, "The element width should be 200px");
            var borderImageSource = GetCurrentStyle("border-image-source");
            prop = borderImageSource.indexOf("none")!=-1;
            assert_true(prop, "The element border-image-source should be none");
            var borderImageWidth = GetCurrentStyle("border-image-width");
            prop = borderImageWidth.indexOf("50px")!=-1;
            assert_true(prop, "The element border-image-width should be 50px");
        });
        t.done();
    </script>
</body>
